+++ examples/react/src/config.ts
@@ -27,7 +27,8 @@ export const connectConfig: ConnectConfig = {
   defaultTheme: 'dark',
   signIn: {
     projectName: 'Sequence Web SDK Demo',
-    useMock: isDebugMode
+    useMock: isDebugMode,
+    showWalletAuthOptionsFirst: false
   },
   // Custom css injected into shadow dom
   // customCSS: `
@@ -183,10 +184,6 @@ export const getErc1155SaleContractConfig = (walletAddress: string) => ({
 export const checkoutConfig: SequenceCheckoutConfig = {
   env: isDev
     ? {
-        sardineCheckoutUrl: 'https://sardine-checkout-sandbox.sequence.info',
-        sardineOnRampUrl: 'https://crypto.sandbox.sardine.ai/',
-        transakApiUrl: 'https://global-stg.transak.com',
-        transakApiKey: 'c20f2a0e-fe6a-4133-8fa7-77e9f84edf98',
         forteWidgetUrl: 'https://payments.sandbox.lemmax.com/forte-payments-widget.js'
       }
     : undefined