+++ packages/checkout/src/views/PaymentSelection/PayWithCreditCard/index.tsx
@@ -14,7 +14,7 @@ interface PayWithCreditCardProps {
   skipOnCloseCallback: () => void
 }
 
-type BasePaymentProviderOptions = 'sardine' | 'transak' | 'forte'
+type BasePaymentProviderOptions = 'transak' | 'forte'
 type CustomPaymentProviderOptions = 'custom'
 type PaymentProviderOptions = BasePaymentProviderOptions | CustomPaymentProviderOptions
 
@@ -27,7 +27,6 @@ export const PayWithCreditCard = ({ settings, disableButtons, skipOnCloseCallbac
     txData,
     collectibles,
     collectionAddress,
-    sardineConfig,
     onSuccess = () => {},
     onError = () => {},
     onClose = () => {},
@@ -63,7 +62,6 @@ export const PayWithCreditCard = ({ settings, disableButtons, skipOnCloseCallbac
           onClickCustomProvider()
         }
         return
-      case 'sardine':
       case 'transak':
       case 'forte':
         onPurchase()
@@ -108,7 +106,7 @@ export const PayWithCreditCard = ({ settings, disableButtons, skipOnCloseCallbac
         nftDecimals: collectible.decimals === undefined ? undefined : String(collectible.decimals),
         provider: selectedPaymentProvider as BasePaymentProviderOptions,
         calldata: txData,
-        approvedSpenderAddress: sardineConfig?.approvedSpenderAddress || settings.approvedSpenderAddress,
+        approvedSpenderAddress: settings.approvedSpenderAddress,
         supplementaryAnalyticsInfo,
         transakConfig,
         forteConfig
@@ -123,46 +121,35 @@ export const PayWithCreditCard = ({ settings, disableButtons, skipOnCloseCallbac
   const Options = () => {
     return (
       <div className="flex flex-col justify-center items-center gap-2 w-full">
-        {/* Only 1 option will be displayed, even if multiple providers are passed */}
-        {creditCardProviders
-          .slice(0, 1)
-          .filter(provider => {
-            // cannot display transak checkout if the settings aren't provided
-            if (provider === 'transak' && !settings.transakConfig) {
-              return false
-            }
-            return true
-          })
-          .map(creditCardProvider => {
-            switch (creditCardProvider) {
-              case 'sardine':
-              case 'transak':
-              case 'forte':
-              case 'custom':
-                return (
-                  <Card
-                    className="flex justify-between items-center p-4 cursor-pointer"
-                    key="sardine"
-                    onClick={() => {
-                      setSelectedPaymentProvider(creditCardProvider)
-                    }}
-                    disabled={disableButtons}
-                  >
-                    <div className="flex flex-row gap-3 items-center">
-                      <PaymentsIcon className="text-white" />
-                      <Text color="primary" variant="normal" fontWeight="bold">
-                        Pay with credit or debit card
-                      </Text>
-                    </div>
-                    <div style={{ transform: 'rotate(-45deg)' }}>
-                      <ArrowRightIcon className="text-white" />
-                    </div>
-                  </Card>
-                )
-              default:
-                return null
-            }
-          })}
+        {creditCardProviders.slice(0, 1).map(creditCardProvider => {
+          switch (creditCardProvider) {
+            case 'transak':
+            case 'forte':
+            case 'custom':
+              return (
+                <Card
+                  className="flex justify-between items-center p-4 cursor-pointer"
+                  key="sardine"
+                  onClick={() => {
+                    setSelectedPaymentProvider(creditCardProvider)
+                  }}
+                  disabled={disableButtons}
+                >
+                  <div className="flex flex-row gap-3 items-center">
+                    <PaymentsIcon className="text-white" />
+                    <Text color="primary" variant="normal" fontWeight="bold">
+                      Pay with credit or debit card
+                    </Text>
+                  </div>
+                  <div style={{ transform: 'rotate(-45deg)' }}>
+                    <ArrowRightIcon className="text-white" />
+                  </div>
+                </Card>
+              )
+            default:
+              return null
+          }
+        })}
       </div>
     )
   }